server/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
asyncpg==0.29.0

# Authentication & Security
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
itsdangerous==2.1.2

# OAuth
authlib==1.3.0
httpx==0.26.0

# Rate Limiting
slowapi==0.1.9

# Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# CORS (built into FastAPI/Starlette, no extra package needed)
# python-cors removed - CORS is handled by FastAPI's CORSMiddleware

# Testing (pytest-asyncio>=0.24 required for pytest 8+ collection)
pytest>=8.0.0,<10
pytest-asyncio>=0.24.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
